pyproject.toml
[project]
name = "hera-video-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "google-genai>=1.12.1",
    "imgcat>=0.6.0",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "yt-dlp>=2025.3.31",
]
